requirement.txt
gradio>=4.0.0
python-dotenv>=1.0.0
qdrant-client>=1.9.0
camel-ai>=0.1.0
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0